projeto_sql/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
mysql-connector-python>=8.0.0
openpyxl>=3.0.0